Feat/woz assign

+++ apps/builder/components/shared/Graph/Nodes/ItemNode/ItemNodeContent/contents/ButtonNodeContent.tsx
@@ -6,7 +6,8 @@ import {
   IconButton,
   Flex,
 } from '@chakra-ui/react'
-import { PlusIcon } from 'assets/icons'
+import { PlusIcon, TrashIcon } from 'assets/icons'
+import { TextBox } from 'components/shared/Textbox/TextBox'
 import { useTypebot } from 'contexts/TypebotContext'
 import { ButtonItem, ItemIndices, ItemType } from 'models'
 import React, { useEffect, useRef, useState } from 'react'
@@ -38,8 +39,14 @@ export const ButtonNodeContent = ({ item, indices, isMouseOver }: Props) => {
       updateItem(indices, { content: itemValue === '' ? undefined : itemValue })
   }
 
+  const hasMoreThanOneItem = () => indices.itemsCount && indices.itemsCount > 1
+
+  const isReadOnly = () => {
+    return !!item.readonly
+  }
+
   const handleKeyPress = (e: React.KeyboardEvent<HTMLDivElement>) => {
-    if (e.key === 'Escape' && itemValue === 'Editar opção de resposta')
+    if (e.key === 'Escape' && itemValue === 'Editar opção de resposta' && hasMoreThanOneItem())
       deleteItem(indices)
     if (e.key === 'Enter' && itemValue !== '' && initialContent === '')
       handlePlusClick()
@@ -51,6 +58,10 @@ export const ButtonNodeContent = ({ item, indices, isMouseOver }: Props) => {
       { stepId: item.stepId, type: ItemType.BUTTON },
       { ...indices, itemIndex }
     )
+  }
+
+  const handleDeleteClick = () => {
+    deleteItem(indices)
   }
 
   return (
@@ -74,7 +85,7 @@ export const ButtonNodeContent = ({ item, indices, isMouseOver }: Props) => {
           px={4}
           py={2}
         />
-        <EditableInput px={4} py={2} />
+        <EditableInput px={4} py={2} readOnly={isReadOnly()} />
       </Editable>
       <Fade
         in={isMouseOver}
@@ -94,6 +105,15 @@ export const ButtonNodeContent = ({ item, indices, isMouseOver }: Props) => {
           colorScheme="gray"
           onClick={handlePlusClick}
         />
+        {hasMoreThanOneItem() && !isReadOnly() && (
+          <IconButton
+            aria-label="Delete item"
+            icon={<TrashIcon />}
+            size="xs"
+            shadow="md"
+            colorScheme="gray"
+            onClick={handleDeleteClick}
+          />)}
       </Fade>
     </Flex>
   )
